Refactor API structure
