+++ go.mod
@@ -2,21 +2,18 @@ module github.com/tidepool-org/tide-whisperer
 
 go 1.12
 
-replace github.com/tidepool-org/shoreline => ./
-
 replace github.com/tidepool-org/go-common => github.com/mdblp/go-common v0.3.0
 
 require (
 	github.com/DataDog/zstd v1.4.4 // indirect
-	github.com/daaku/go.httpgzip v0.0.0-20180202095102-86d27ccd810b
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/go-stack/stack v1.8.0 // indirect
 	github.com/golang/snappy v0.0.1 // indirect
 	github.com/google/go-cmp v0.3.1 // indirect
 	github.com/google/uuid v1.1.1
-	github.com/gorilla/context v1.1.1 // indirect
-	github.com/gorilla/mux v1.7.3 // indirect
-	github.com/gorilla/pat v1.0.1
+	github.com/gorilla/handlers v1.4.2
+	github.com/gorilla/mux v1.7.3
+	github.com/kr/pretty v0.2.0 // indirect
 	github.com/pkg/errors v0.8.1 // indirect
 	github.com/stretchr/testify v1.3.0 // indirect
 	github.com/tidepool-org/go-common v0.0.0-00010101000000-000000000000
