+++ tide-whisperer.go
@@ -1,21 +1,15 @@
 package main
 
 import (
-	"crypto/rand"
 	"crypto/tls"
-	"encoding/hex"
-	"encoding/json"
-	"fmt"
 	"log"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 
-	httpgzip "github.com/daaku/go.httpgzip"
-	"github.com/google/uuid"
-	"github.com/gorilla/pat"
+	"github.com/gorilla/handlers"
+	"github.com/gorilla/mux"
 
 	common "github.com/tidepool-org/go-common"
 	"github.com/tidepool-org/go-common/clients"
@@ -23,9 +17,8 @@ import (
 	"github.com/tidepool-org/go-common/clients/hakken"
 	"github.com/tidepool-org/go-common/clients/mongo"
 	"github.com/tidepool-org/go-common/clients/shoreline"
-	"github.com/tidepool-org/go-common/clients/status"
-
 	"github.com/tidepool-org/tide-whisperer/auth"
+	"github.com/tidepool-org/tide-whisperer/data"
 	"github.com/tidepool-org/tide-whisperer/store"
 )
 
@@ -38,54 +31,19 @@ type (
 		Mongo               mongo.Config        `json:"mongo"`
 		store.SchemaVersion `json:"schemaVersion"`
 	}
-
-	// so we can wrap and marshal the detailed error
-	detailedError struct {
-		Status int `json:"status"`
-		//provided to user so that we can better track down issues
-		ID              string `json:"id"`
-		Code            string `json:"code"`
-		Message         string `json:"message"`
-		InternalMessage string `json:"-"` //used only for logging so we don't want to serialize it out
-	}
-	//generic type as device data can be comprised of many things
-	deviceData map[string]interface{}
 )
-
-var (
-	errorStatusCheck       = detailedError{Status: http.StatusInternalServerError, Code: "data_status_check", Message: "checking of the status endpoint showed an error"}
-	errorNoViewPermission  = detailedError{Status: http.StatusForbidden, Code: "data_cant_view", Message: "user is not authorized to view data"}
-	errorNoPermissions     = detailedError{Status: http.StatusInternalServerError, Code: "data_perms_error", Message: "error finding permissions for user"}
-	errorRunningQuery      = detailedError{Status: http.StatusInternalServerError, Code: "data_store_error", Message: "internal server error"}
-	errorLoadingEvents     = detailedError{Status: http.StatusInternalServerError, Code: "data_marshal_error", Message: "internal server error"}
-	errorInvalidParameters = detailedError{Status: http.StatusInternalServerError, Code: "invalid_parameters", Message: "one or more parameters are invalid"}
-
-	storage store.Storage
-)
-
-const (
-	dataAPIPrefix             = "api/data "
-	medtronicLoopBoundaryDate = "2017-09-01"
-	slowQueryDuration         = 0.1 // seconds
-)
-
-//set the intenal message that we will use for logging
-func (d detailedError) setInternalMessage(internal error) detailedError {
-	d.InternalMessage = internal.Error()
-	return d
-}
 
 func main() {
 	var config Config
 
-	log.SetPrefix(dataAPIPrefix)
+	log.SetPrefix(data.DATA_API_PREFIX)
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
 
 	if err := common.LoadEnvironmentConfig(
 		[]string{"TIDEPOOL_TIDE_WHISPERER_SERVICE", "TIDEPOOL_TIDE_WHISPERER_ENV"},
 		&config,
 	); err != nil {
-		log.Fatal(dataAPIPrefix, "Problem loading config: ", err)
+		log.Fatal("Problem loading config: ", err)
 	}
 
 	// server secret may be passed via a separate env variable to accommodate easy secrets injection via Kubernetes
@@ -107,7 +65,7 @@ func main() {
 
 	authClient, err := auth.NewClient(config.Auth, httpClient)
 	if err != nil {
-		log.Fatal(dataAPIPrefix, err)
+		log.Fatal(err)
 	}
 
 	hakkenClient := hakken.NewHakkenBuilder().
@@ -116,11 +74,11 @@ func main() {
 
 	if !config.HakkenConfig.SkipHakken {
 		if err := hakkenClient.Start(); err != nil {
-			log.Fatal(dataAPIPrefix, err)
+			log.Fatal(err)
 		}
 		defer func() {
 			if err := hakkenClient.Close(); err != nil {
-				log.Panic(dataAPIPrefix, "Error closing hakkenClient, panicing to get stacks: ", err)
+				log.Panic("Error closing hakkenClient, panicing to get stacks: ", err)
 			}
 		}()
 	} else {
@@ -133,273 +91,34 @@ func main() {
 		WithConfig(&config.ShorelineConfig.ShorelineClientConfig).
 		Build()
 
-	gatekeeperClient := clients.NewGatekeeperClientBuilder().
+	permsClient := clients.NewGatekeeperClientBuilder().
 		WithHostGetter(config.GatekeeperConfig.ToHostGetter(hakkenClient)).
 		WithHttpClient(httpClient).
 		WithTokenProvider(shorelineClient).
 		Build()
 
-	userCanViewData := func(authenticatedUserID string, targetUserID string) bool {
-		if authenticatedUserID == targetUserID {
-			return true
-		}
-
-		perms, err := gatekeeperClient.UserInGroup(authenticatedUserID, targetUserID)
-		if err != nil {
-			log.Println(dataAPIPrefix, "Error looking up user in group", err)
-			return false
-		}
-
-		log.Println(perms)
-		return !(perms["root"] == nil && perms["view"] == nil)
-	}
-
-	logError := func(err *detailedError, startedAt time.Time) {
-		err.ID = uuid.New().String()
-		log.Println(dataAPIPrefix, fmt.Sprintf("[%s][%s] failed after [%.3f]secs with error [%s][%s] ", err.ID, err.Code, time.Now().Sub(startedAt).Seconds(), err.Message, err.InternalMessage))
-	}
-	//log error detail and write as application/json
-	jsonError := func(res http.ResponseWriter, err detailedError, startedAt time.Time) {
-		logError(&err, startedAt)
-		jsonErr, _ := json.Marshal(err)
-
-		res.Header().Add("content-type", "application/json")
-		res.WriteHeader(err.Status)
-		res.Write(jsonErr)
-	}
-
 	if err := shorelineClient.Start(); err != nil {
 		log.Fatal(err)
 	}
 
-	storage := store.NewMongoStoreClient(&config.Mongo)
-	defer storage.Disconnect()
-	storage.EnsureIndexes()
-
-	router := pat.New()
+	rtr := mux.NewRouter()
 
 	/*
-	 Gloo performs autodiscovery by trying certain paths,
-	 including /swagger, /v1, and v2.  Unfortunately, tide-whisperer
-	 interprets those paths as userids.  To avoid misleading
-	 error messages, we catch these calls and return an error
-	 code.
-	*/
-	router.Add("GET", "/swagger", http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		res.WriteHeader(501)
-		return
-	}))
+	 * Data-Api setup
+	 */
 
-	router.Add("GET", "/v1", http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		res.WriteHeader(501)
-		return
-	}))
+	dataapi := data.InitApi(config.Mongo, shorelineClient, authClient, permsClient, config.SchemaVersion)
+	dataapi.SetHandlers("", rtr)
 
-	router.Add("GET", "/v2", http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		res.WriteHeader(501)
-		return
-	}))
-
-	router.Add("GET", "/status", http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		start := time.Now()
-		var s status.ApiStatus
-		if err := storage.Ping(); err != nil {
-			errorLog := errorStatusCheck.setInternalMessage(err)
-			logError(&errorLog, start)
-			s = status.NewApiStatus(errorLog.Status, err.Error())
-		} else {
-			s = status.NewApiStatus(http.StatusOK, "OK")
-		}
-		if jsonDetails, err := json.Marshal(s); err != nil {
-			jsonError(res, errorLoadingEvents.setInternalMessage(err), start)
-		} else {
-			res.Header().Add("content-type", "application/json")
-			res.WriteHeader(s.Status.Code)
-			res.Write(jsonDetails)
-		}
-		return
-	}))
-
-	// The /data/userId endpoint retrieves device/health data for a user based on a set of parameters
-	// userid: the ID of the user you want to retrieve data for
-	// uploadId (optional) : Search for Tidepool data by uploadId. Only objects with a uploadId field matching the specified uploadId param will be returned.
-	// deviceId (optional) : Search for Tidepool data by deviceId. Only objects with a deviceId field matching the specified uploadId param will be returned.
-	// type (optional) : The Tidepool data type to search for. Only objects with a type field matching the specified type param will be returned.
-	//					can be /userid?type=smbg or a comma seperated list e.g /userid?type=smgb,cbg . If is a comma seperated
-	//					list, then objects matching any of the sub types will be returned
-	// subType (optional) : The Tidepool data subtype to search for. Only objects with a subtype field matching the specified subtype param will be returned.
-	//					can be /userid?subtype=physicalactivity or a comma seperated list e.g /userid?subtypetype=physicalactivity,steps . If is a comma seperated
-	//					list, then objects matching any of the types will be returned
-	// startDate (optional) : Only objects with 'time' field equal to or greater than start date will be returned.
-	//					Must be in ISO date/time format e.g. 2015-10-10T15:00:00.000Z
-	// endDate (optional) : Only objects with 'time' field less than to or equal to start date will be returned.
-	//					Must be in ISO date/time format e.g. 2015-10-10T15:00:00.000Z
-	// latest (optional) : Returns only the most recent results for each `type` matching the results filtered by the other query parameters
-	router.Add("GET", "/{userID}", httpgzip.NewHandler(http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
-		start := time.Now()
-
-		storageWithCtx := storage.WithContext(req.Context())
-
-		queryParams, err := store.GetParams(req.URL.Query(), &config.SchemaVersion, &config.Mongo)
-
-		if err != nil {
-			log.Println(dataAPIPrefix, fmt.Sprintf("Error parsing query params: %s", err))
-			jsonError(res, errorInvalidParameters, start)
-			return
-		}
-
-		var td *shoreline.TokenData
-		if sessionToken := req.Header.Get("x-tidepool-session-token"); sessionToken != "" {
-			td = shorelineClient.CheckToken(sessionToken)
-		} else if restrictedTokens, found := req.URL.Query()["restricted_token"]; found && len(restrictedTokens) == 1 {
-			restrictedToken, restrictedTokenErr := authClient.GetRestrictedToken(req.Context(), restrictedTokens[0])
-			if restrictedTokenErr == nil && restrictedToken != nil && restrictedToken.Authenticates(req) {
-				td = &shoreline.TokenData{UserID: restrictedToken.UserID}
-			}
-		}
-
-		userID := queryParams.UserID
-		if td == nil || !(td.IsServer || td.UserID == userID || userCanViewData(td.UserID, userID)) {
-			log.Printf("userid %v", userID)
-			jsonError(res, errorNoViewPermission, start)
-			return
-		}
-
-		requestID := NewRequestID()
-		queryStart := time.Now()
-		if _, ok := req.URL.Query()["carelink"]; !ok {
-			if hasMedtronicDirectData, medtronicErr := storageWithCtx.HasMedtronicDirectData(queryParams.UserID); medtronicErr != nil {
-				log.Printf("%s request %s user %s HasMedtronicDirectData returned error: %s", dataAPIPrefix, requestID, userID, medtronicErr)
-				jsonError(res, errorRunningQuery, start)
-				return
-			} else if !hasMedtronicDirectData {
-				queryParams.Carelink = true
-			}
-			if queryDuration := time.Now().Sub(queryStart).Seconds(); queryDuration > slowQueryDuration {
-				// XXX replace with metrics
-				//log.Printf("%s request %s user %s HasMedtronicDirectData took %.3fs", DATA_API_PREFIX, requestID, userID, queryDuration)
-			}
-			queryStart = time.Now()
-		}
-		if !queryParams.Dexcom {
-			dexcomDataSource, dexcomErr := storageWithCtx.GetDexcomDataSource(queryParams.UserID)
-			if dexcomErr != nil {
-				log.Printf("%s request %s user %s GetDexcomDataSource returned error: %s", dataAPIPrefix, requestID, userID, dexcomErr)
-				jsonError(res, errorRunningQuery, start)
-				return
-			}
-			queryParams.DexcomDataSource = dexcomDataSource
-
-			if queryDuration := time.Now().Sub(queryStart).Seconds(); queryDuration > slowQueryDuration {
-				log.Printf("%s request %s user %s GetDexcomDataSource took %.3fs", dataAPIPrefix, requestID, userID, queryDuration)
-			}
-			queryStart = time.Now()
-		}
-		if _, ok := req.URL.Query()["medtronic"]; !ok {
-			hasMedtronicLoopData, medtronicErr := storageWithCtx.HasMedtronicLoopDataAfter(queryParams.UserID, medtronicLoopBoundaryDate)
-			if medtronicErr != nil {
-				log.Printf("%s request %s user %s HasMedtronicLoopDataAfter returned error: %s", dataAPIPrefix, requestID, userID, medtronicErr)
-				jsonError(res, errorRunningQuery, start)
-				return
-			}
-			if !hasMedtronicLoopData {
-				queryParams.Medtronic = true
-			}
-			if queryDuration := time.Now().Sub(queryStart).Seconds(); queryDuration > slowQueryDuration {
-				log.Printf("%s request %s user %s HasMedtronicLoopDataAfter took %.3fs", dataAPIPrefix, requestID, userID, queryDuration)
-			}
-			queryStart = time.Now()
-		}
-		if !queryParams.Medtronic {
-			medtronicUploadIds, medtronicErr := storageWithCtx.GetLoopableMedtronicDirectUploadIdsAfter(queryParams.UserID, medtronicLoopBoundaryDate)
-			if medtronicErr != nil {
-				log.Printf("%s request %s user %s GetLoopableMedtronicDirectUploadIdsAfter returned error: %s", dataAPIPrefix, requestID, userID, medtronicErr)
-				jsonError(res, errorRunningQuery, start)
-				return
-			}
-			queryParams.MedtronicDate = medtronicLoopBoundaryDate
-			queryParams.MedtronicUploadIds = medtronicUploadIds
-
-			if queryDuration := time.Now().Sub(queryStart).Seconds(); queryDuration > slowQueryDuration {
-				// XXX replace with metrics
-				//log.Printf("%s request %s user %s GetLoopableMedtronicDirectUploadIdsAfter took %.3fs", DATA_API_PREFIX, requestID, userID, queryDuration)
-			}
-			queryStart = time.Now()
-		}
-
-		iter, err := storageWithCtx.GetDeviceData(queryParams)
-		if err != nil {
-			log.Printf("%s request %s user %s Mongo Query returned error: %s", dataAPIPrefix, requestID, userID, err)
-		}
-
-		defer iter.Close(req.Context())
-
-		var parametersHistory map[string]interface{}
-		var parametersHistoryErr error
-		if store.InArray("pumpSettings", queryParams.Types) || (len(queryParams.Types) == 1 && queryParams.Types[0] == "") {
-			log.Printf("Calling GetDiabeloopParametersHistory")
-
-			if parametersHistory, parametersHistoryErr = storage.GetDiabeloopParametersHistory(queryParams.UserID, queryParams.LevelFilter); parametersHistoryErr != nil {
-				log.Printf("%s request %s user %s GetDiabeloopParametersHistory returned error: %s", dataAPIPrefix, requestID, userID, parametersHistoryErr)
-				jsonError(res, errorRunningQuery, start)
-				return
-			}
-		}
-		var writeCount int
-
-		res.Header().Add("Content-Type", "application/json")
-
-		res.Write([]byte("["))
-
-		for iter.Next(req.Context()) {
-			var results map[string]interface{}
-			err := iter.Decode(&results)
-			if err != nil {
-				log.Printf("%s request %s user %s Mongo Decode returned error: %s", dataAPIPrefix, requestID, userID, err)
-			}
-
-			if queryParams.Latest {
-				// If we're using the `latest` parameter, then we ran an `$aggregate` query to get only the latest data.
-				// Since we use Mongo 3.2, we can't use the $replaceRoot function, so we need to manually extract the
-				// latest subdocument here. When we move to MongoDB 3.4+ and can use $replaceRoot, we can get rid of this
-				// conditional block. We'd also need to fix the corresponding code in `store.go`
-				results = results["latest_doc"].(map[string]interface{})
-			}
-			if len(results) > 0 {
-				if results["type"].(string) == "pumpSettings" && parametersHistory != nil {
-					payload := results["payload"].(map[string]interface{})
-					payload["history"] = parametersHistory["history"]
-					results["payload"] = payload
-				}
-				if bytes, err := json.Marshal(results); err != nil {
-					log.Printf("%s request %s user %s Marshal returned error: %s", dataAPIPrefix, requestID, userID, err)
-				} else {
-					if writeCount > 0 {
-						res.Write([]byte(","))
-					}
-					res.Write([]byte("\n"))
-					res.Write(bytes)
-					writeCount++
-				}
-			}
-		}
-
-		if writeCount > 0 {
-			res.Write([]byte("\n"))
-		}
-		res.Write([]byte("]"))
-
-		if queryDuration := time.Now().Sub(queryStart).Seconds(); queryDuration > slowQueryDuration {
-			// XXX use metrics
-			//log.Printf("%s request %s user %s GetDeviceData took %.3fs", DATA_API_PREFIX, requestID, userID, queryDuration)
-		}
-		log.Printf("%s request %s user %s took %.3fs returned %d records", dataAPIPrefix, requestID, userID, time.Now().Sub(start).Seconds(), writeCount)
-	})))
+	// ability to return compressed (gzip/deflate) responses if client browser accepts it
+	// this is interesting to minimise network traffic especially if we expect to have long
+	// responses such as what the GetData() route here can return
+	gzipHandler := handlers.CompressHandler(rtr)
 
 	done := make(chan bool)
 	server := common.NewServer(&http.Server{
 		Addr:    config.Service.GetPort(),
-		Handler: router,
+		Handler: gzipHandler,
 	})
 
 	var start func() error
@@ -410,7 +129,7 @@ func main() {
 		start = func() error { return server.ListenAndServe() }
 	}
 	if err := start(); err != nil {
-		log.Fatal(dataAPIPrefix, err)
+		log.Fatal(err)
 	}
 	hakkenClient.Publish(&config.Service)
 
@@ -419,7 +138,7 @@ func main() {
 	go func() {
 		for {
 			sig := <-signals
-			log.Printf(dataAPIPrefix+" Got signal [%s]", sig)
+			log.Printf("Got signal [%s]", sig)
 
 			if sig == syscall.SIGINT || sig == syscall.SIGTERM {
 				server.Close()
@@ -429,11 +148,4 @@ func main() {
 	}()
 
 	<-done
-}
-
-// NewRequestID returns a new random hexadecimal ID
-func NewRequestID() string {
-	bytes := make([]byte, 8)
-	_, _ = rand.Read(bytes) // In case of failure, do not fail request, just use default bytes (zero)
-	return hex.EncodeToString(bytes)
 }
